Add CreateElement module
